#CT-70 #CT-54 Add AuthService for user authentication and session management

+++ lib/screens/add_appointment_screen.dart
@@ -145,7 +145,7 @@ class _AddAppointmentScreenState extends State<AddAppointmentScreen> {
             .toIso8601String(),
         'description': descriptionController.text,
         'price': double.tryParse(priceController.text) ?? 0.0,
-        'user_percentage': int.tryParse(userPercentController.text) ?? 100,
+        'user_percentage': double.tryParse(userPercentController.text) ?? 100.0,
         'has_extra_cost': hasExtraCost,
         'extra_cost': hasExtraCost
             ? (double.tryParse(extraCostController.text) ?? 0.0)
